simiple stories cache and removed active clients
Added simple stories cache as we can assume that the story did not
changed much in a span of few seconds.  We still always query the
list of the best stories.

Also more tests revealed unstable nature of using many clients.  Using
two clients does not yield the fastest results but these are the most
stable - typically around 1-2 seconds.  Where using 50-100 clients
typically produce results under 1 second but sometimes will stale the
connection for half a minute or longer.  I prefer more stable but
slightly slower solution as with the cache it is now fast enough.