unit tests
Changed few file names and the MockData setting to StaticData.
A solution file was added.

Changed how stories are cached.  Now stories do not have Id property
(as it was not deserialized from JSON anyway) and the DateTime.UtcNow
is individual to each fetched story instead of reusing the fetchTime
variable which resulted in all bag of stories having exactly the same
cache time - which was both not true and resulted in worst distribution
and cache reuse as every five seconds all the items in the cache would
be invalid instead of only couple individual ones.

Also the GetNBestStoriesIdsAsync has been changed to
GetAllBestStoriesIdsAsync as that's the only use case
and I don't want to support filter parameters which are not used.